Move queues file into separate folder
